cloud_function/requirements.txt
google-cloud-storage
google-cloud-bigquery
pandas
gcsfs
pyarrow